refactor: #184 merged BlueprintResolver into SchemaCatalog
