+++ packages/core/src/lib/entity/blueprint/index.ts
@@ -1,3 +1,3 @@
-export * from "./blueprint-resolver";
+// export * from "./blueprint-resolver";
 export * from "./instance";
 export * from "./property";
