+++ packages/examples/apps/music-box-api/src/app/controllers/artists.controller.ts
@@ -1,12 +1,12 @@
-import { BlueprintResolver, IEntitySchema, Query as EntityQuery } from "@entity-space/core";
+import { IEntitySchema, Query as EntityQuery, SchemaCatalog } from "@entity-space/core";
 import { Artist, ArtistBlueprint } from "@entity-space/examples/libs/music-model";
 import { Body, Controller, Get, Param, ParseIntPipe, Patch, Post } from "@nestjs/common";
 import { DiskDbService } from "../disk-db.service";
 
 @Controller("artists")
 export class ArtistsController {
-    constructor(private readonly diskDbService: DiskDbService, private readonly blueprintResolver: BlueprintResolver) {
-        this.schema = this.blueprintResolver.resolve(ArtistBlueprint);
+    constructor(private readonly diskDbService: DiskDbService, private readonly schemas: SchemaCatalog) {
+        this.schema = this.schemas.resolve(ArtistBlueprint);
     }
 
     private readonly schema: IEntitySchema;
